app: CLI: Pico-SDK "advanced UART" sample code working, second uart on RPi Pico board now talking bi-directionally.
Code of this commit is a chimera of sorts, especially with a need to connect UART1_IRQ hardware interrupt via a Zephyr API in order for an interrupt which Pico SDK enables to be properly mapped in Zephyr scheduler context.  Also "chimera code" in CMakeLists.txt in effort to express needed library paths to Pico SDK headers.

On "To Do" list is to figure out whether there is a clean, correct way in Zephyr 3.3.0 to enable Pico-SDK and Pico Examples.  The SDK ispresent along side Zephyr source tree in `modules/hal/rpi_pico/`.  Pico samples are not present in Zephyr 3.3.0 manifest provided projects.